requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
langchain-cohere
langchain-text-splitters
faiss-cpu
pydantic
